20251224 - Abandoned Google Gemini LLM
Unfortunately I deprecated the use of the Google Gemini  LLM components as their free tier limits have been reduced to an amount that is not practical for further development in this project. The code is still available but commented out in case you would like to use one of the paid tiers.